Restructure frontend entry files into .frontend/ to match Studio conventions
Move App.tsx, index.tsx, index.html, and index.css from the project root
into .frontend/ so that Vite's root points to .frontend/ itself, matching
the layout used by all Studio frontend modules.

- Fix .frontend/tsconfig.json extends path: ../../ → ../../../ so the
  base config (jsx, noEmit, strict, etc.) is correctly inherited
- Fix .frontend/vite.config.ts root: '../' → './' and vitest setupFiles
  path to correctly resolve Samples/.frontend/vitest.setup.ts
- Simplify root tsconfig.json to just extend .frontend/tsconfig.json
- Add @cratis/components ^1.10.2 dependency to both packages so subpath
  imports (@cratis/components/Toolbar, /CommandDialog, etc.) resolve
- Fix ToolbarButton tooltip → title prop to match current API
- Fix AddBook.tsx: update AllAuthors import to ObserveAll proxy
