+++ Library/Lending/.frontend/vite.config.ts
@@ -50,7 +50,7 @@ export default defineConfig({
         },
         exclude: ['../dist/**', '../node_modules/**', 'node_modules/**', '../wwwroot/**', 'wwwroot/**', '../**/given/**'],
         include: ['../**/for_*/when_*/**/*.ts', '../**/for_*/**/when_*.ts'],
-        setupFiles: fileURLToPath(new URL('../../.frontend/vitest.setup.ts', import.meta.url))
+        setupFiles: fileURLToPath(new URL('../../../.frontend/vitest.setup.ts', import.meta.url))
     },
     plugins: [
         react(),
