+++ Library/Lending/package.json
@@ -20,6 +20,7 @@
     "@cratis/arc.react": "20.39.1",
     "@cratis/arc.react.mvvm": "20.39.1",
     "@cratis/arc.vite": "20.39.1",
+    "@cratis/components": "^1.10.2",
     "allotment": "1.20.5",
     "echarts": "6.0.0",
     "mobx": "6.15.3",
