+++ Library/Members/.frontend/index.tsx
@@ -9,7 +9,7 @@ import './index.css';
 import React from 'react';
 import App from "./App";
 import { configure as configureMobx } from 'mobx';
-import { Bindings } from './Bindings';
+import { Bindings } from '../Bindings';
 
 Bindings.initialize();
